Move tllog models to tllog directory
